pyproject.toml
[project]
name = "vibe-coding-tools"
version = "0.1.0"
description = "AI Coding Rules automation tools"
requires-python = ">=3.10"
dependencies = [
    "pre-commit>=3.0.0",
    "pyyaml>=6.0.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Ruff 配置（与 IDE 中使用的 Ruff 扩展保持一致）
[tool.ruff]
# 启用推荐的规则集
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

# 行长度限制
line-length = 88

# 排除的目录
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "node_modules",
    "build",
    "dist",
]

[tool.ruff.format]
# 使用双引号
quote-style = "double"
# 缩进使用空格
indent-style = "space"
# 行尾使用 LF
line-ending = "lf"

